+++ src/types.ts
@@ -1233,7 +1233,10 @@ export type OptionalHandler<
 ) => MaybePromise<
 	{} extends Route['response']
 		? unknown
-		: Route['response'][keyof Route['response']] | void
+		:
+				| Route['response'][keyof Route['response']]
+				| InlineHandlerResponse<Route['response']>
+				| void
 >
 
 export type AfterHandler<
@@ -1370,7 +1373,11 @@ export type PreHandler<
 		derive: {}
 		resolve: {}
 	}
-> = (context: PreContext<Singleton>) => MaybePromise<Route['response'] | void>
+> = (
+	context: PreContext<Singleton>
+) => MaybePromise<
+	Route['response'] | InlineHandlerResponse<Route['response']> | void
+>
 
 export type AfterResponseHandler<
 	in out Route extends RouteSchema = {},
@@ -1390,7 +1397,9 @@ export type AfterResponseHandler<
 		 */
 		response: {} extends Route['response']
 			? unknown
-			: Route['response'][keyof Route['response']]
+			:
+					| Route['response'][keyof Route['response']]
+					| InlineHandlerResponse<Route['response']>
 	}
 ) => MaybePromise<unknown>